requirements.txt
python-twitter==3.5
secure-smtplib==0.1.1
oauthlib==3.1.1
requests-oauthlib==1.3.0
selenium==3.141.0